重构规范化微信支付相关接口类的命名
